+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Portuguese install instructions mentioning update procedure.
   * English install instructions.
 * Fixed bugs:
+  * When hidden one of the background images in cave.th2 (demo.zip) the other image moves.
 * Infrastructure maintenance:
 
 ## 0.2.31 - 2026-01-15 - The [EGB](https://www.egb.org.br) release